Move translateops functions to separate file
