+++ src/ngraph_builder.cc
@@ -2696,13 +2696,12 @@ Status Builder::TranslateGraph(
 
   for (const auto n : ordered) {
     cout << "XXXXX " << n->type_string() << "\n";
-    if (n->type_string() == "Const"){
+    if (n->type_string() == "Const") {
       Tensor result;
       vector<int64> vect;
       result.FromProto(n->def().attr().at("value").tensor());
       TensorDataToVector(result, &vect);
-      for (auto vv : vect)
-        cout << " " << vv ;
+      for (auto vv : vect) cout << " " << vv;
       cout << "\n";
     }
     if (n->IsSink() || n->IsSource()) {
