+++ test/test_math_ops.cpp
@@ -14,29 +14,29 @@
  * limitations under the License.
  *******************************************************************************/
 #include "gtest/gtest.h"
-#include "opexecuter.h"
-#include "test_utilities.h"
 #include "ngraph_utils.h"
-#include "tf_graph_writer.h"
+#include "opexecuter.h"
+#include "tensorflow/cc/client/client_session.h"
+#include "tensorflow/cc/ops/standard_ops.h"
 #include "tensorflow/core/common_runtime/dma_helper.h"
 #include "tensorflow/core/framework/graph.pb.h"
 #include "tensorflow/core/framework/op.h"
+#include "tensorflow/core/framework/tensor.h"
 #include "tensorflow/core/framework/tensor_types.h"
 #include "tensorflow/core/graph/algorithm.h"
 #include "tensorflow/core/graph/graph.h"
 #include "tensorflow/core/graph/graph_constructor.h"
 #include "tensorflow/core/platform/env.h"
-#include "tensorflow/cc/client/client_session.h"
-#include "tensorflow/cc/ops/standard_ops.h"
-#include "tensorflow/core/framework/tensor.h"
 #include "tensorflow/core/public/session.h"
- using namespace std;
+#include "test_utilities.h"
+#include "tf_graph_writer.h"
+using namespace std;
 namespace ng = ngraph;
- namespace tensorflow {
- namespace ngraph_bridge {
- namespace testing {
- #define ASSERT_OK(x) ASSERT_EQ((x), ::tensorflow::Status::OK());
- // Test(TestCaseName, TestName)
+namespace tensorflow {
+namespace ngraph_bridge {
+namespace testing {
+#define ASSERT_OK(x) ASSERT_EQ((x), ::tensorflow::Status::OK());
+// Test(TestCaseName, TestName)
 // Please ensure
 // Neither TestCaseName nor TestName should contain underscore
 // https://github.com/google/googletest/blob/master/googletest/docs/primer.md
@@ -46,51 +46,51 @@ TEST(MathOps, Neg) {
   Scope root = Scope::NewRootScope();
   int dim1 = 2;
   int dim2 = 2;
-   Tensor A(DT_FLOAT, TensorShape({dim1, dim2}));
-   AssignInputValues(A, -2.1f);
-   auto R = ops::Neg(root, A);
-   vector<DataType> output_datatypes = {DT_FLOAT};
+  Tensor A(DT_FLOAT, TensorShape({dim1, dim2}));
+  AssignInputValues(A, -2.1f);
+  auto R = ops::Neg(root, A);
+  vector<DataType> output_datatypes = {DT_FLOAT};
   std::vector<Output> sess_run_fetchoutputs = {R};
   OpExecuter opexecuter(root, "Neg", output_datatypes, sess_run_fetchoutputs);
-   opexecuter.ExecuteOnNGraph();
+  opexecuter.ExecuteOnNGraph();
   opexecuter.ExecuteOnTF();
   opexecuter.CompareNgraphAndTF();
 }
 
- TEST(MathOps, Add) {
+TEST(MathOps, Add) {
   // Create a tf graph
   Scope root = Scope::NewRootScope();
   int dim1 = 2;
   int dim2 = 2;
-   Tensor A(DT_FLOAT, TensorShape({dim1, dim2}));
+  Tensor A(DT_FLOAT, TensorShape({dim1, dim2}));
   Tensor B(DT_FLOAT, TensorShape({dim1, dim2}));
-   AssignInputValues(A, 2.1f);
+  AssignInputValues(A, 2.1f);
   AssignInputValues(B, 4.1f);
-   auto R = ops::Add(root, A, B);
-   vector<DataType> output_datatypes = {DT_FLOAT};
+  auto R = ops::Add(root, A, B);
+  vector<DataType> output_datatypes = {DT_FLOAT};
   std::vector<Output> sess_run_fetchoutputs = {R};
   OpExecuter opexecuter(root, "Add", output_datatypes, sess_run_fetchoutputs);
-   opexecuter.ExecuteOnNGraph();
+  opexecuter.ExecuteOnNGraph();
   opexecuter.ExecuteOnTF();
   opexecuter.CompareNgraphAndTF();
 }
- TEST(MathOps, RealDiv) {
+TEST(MathOps, RealDiv) {
   Scope root = Scope::NewRootScope();
   int dim1 = 100;
   int dim2 = 200;
-   Tensor A(DT_FLOAT, TensorShape({dim1, dim2}));
+  Tensor A(DT_FLOAT, TensorShape({dim1, dim2}));
   Tensor B(DT_FLOAT, TensorShape({dim1, dim2}));
-   AssignInputValues(A, 2.0f);
+  AssignInputValues(A, 2.0f);
   AssignInputValues(B, 7.0f);
-   auto R = ops::RealDiv(root, A, B);
-   vector<DataType> output_datatypes = {DT_FLOAT};
-   std::vector<Output> sess_run_fetchoutputs = {R};
+  auto R = ops::RealDiv(root, A, B);
+  vector<DataType> output_datatypes = {DT_FLOAT};
+  std::vector<Output> sess_run_fetchoutputs = {R};
   OpExecuter opexecuter(root, "RealDiv", output_datatypes,
                         sess_run_fetchoutputs);
-   opexecuter.ExecuteOnNGraph();
+  opexecuter.ExecuteOnNGraph();
   opexecuter.ExecuteOnTF();
   opexecuter.CompareNgraphAndTF();
 }
- }  // namespace testing
- }  // namespace ngraph_bridge
+}  // namespace testing
+}  // namespace ngraph_bridge
 }  // namespace tensorflow
