+++ test/test_utilities.cpp
@@ -13,20 +13,20 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *******************************************************************************/
- #include "test_utilities.h"
- using namespace std;
- namespace ng = ngraph;
- namespace tensorflow {
- namespace ngraph_bridge {
- void ActivateNGraph() {
+#include "test_utilities.h"
+using namespace std;
+namespace ng = ngraph;
+namespace tensorflow {
+namespace ngraph_bridge {
+void ActivateNGraph() {
   setenv("NGRAPH_TF_DISABLE_DEASSIGN_CLUSTERS", "1", 1);
   unsetenv("NGRAPH_TF_DISABLE");
 }
- void DeactivateNGraph() {
+void DeactivateNGraph() {
   unsetenv("NGRAPH_TF_DISABLE_DEASSIGN_CLUSTERS");
   setenv("NGRAPH_TF_DISABLE", "1", 1);
 }
- void AssertTensorEquals(Tensor& T1, Tensor& T2) {
+void AssertTensorEquals(Tensor& T1, Tensor& T2) {
   auto T_size = T1.flat<float>().size();
   auto T1_data = T1.flat<float>().data();
   auto T2_data = T2.flat<float>().data();
@@ -36,7 +36,7 @@
     EXPECT_FLOAT_EQ(a, b);
   }
 }
- void AssignInputIntValues(Tensor& A, int maxval) {
+void AssignInputIntValues(Tensor& A, int maxval) {
   auto A_flat = A.flat<int>();
   auto A_flat_data = A_flat.data();
   int counter = 0;
@@ -47,17 +47,17 @@
     }
   }
 }
- void AssignInputValues(Tensor& A, float x) {
+void AssignInputValues(Tensor& A, float x) {
   auto A_flat = A.flat<float>();
   auto A_flat_data = A_flat.data();
   for (int i = 0; i < A_flat.size(); i++) {
     A_flat_data[i] = x;
   }
 }
- void PrintTensor(const Tensor& T1) {
+void PrintTensor(const Tensor& T1) {
   LOG(INFO) << "print tensor values" << T1.DebugString();
 }
- void ValidateTensorData(Tensor& T1, Tensor& T2, float tol) {
+void ValidateTensorData(Tensor& T1, Tensor& T2, float tol) {
   auto T_size = T1.flat<float>().size();
   auto T1_data = T1.flat<float>().data();
   auto T2_data = T2.flat<float>().data();
@@ -73,5 +73,5 @@
     }
   }
 }
- }  // namespace ngraph_bridge
-} //namespace tensorflow
+}  // namespace ngraph_bridge
+}  // namespace tensorflow
